requirements.txt
# requirements.txt
requests
sentence-transformers
faiss-cpu
numpy
fastapi
uvicorn
jinja2
httpx
python-dotenv
transformers
python-multipart
sqlalchemy
aiosqlite
duckduckgo-search
google-api-python-client
duckduckgo-search           
openai           